Moved Attribute into Entity.
